more to setting panel
