+++ src/main/java/ruiseki/okbackpack/common/block/BackpackPanel.java
@@ -1,10 +1,7 @@
 package ruiseki.okbackpack.common.block;
 
-import java.io.File;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 import net.minecraft.client.Minecraft;
@@ -13,14 +10,11 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.EnumChatFormatting;
 
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-import org.lwjgl.input.Keyboard;
 
 import com.cleanroommc.modularui.api.IPanelHandler;
-import com.cleanroommc.modularui.api.UpOrDown;
 import com.cleanroommc.modularui.api.drawable.IKey;
 import com.cleanroommc.modularui.api.widget.IWidget;
 import com.cleanroommc.modularui.api.widget.Interactable;
@@ -39,11 +33,9 @@ import com.cleanroommc.modularui.theme.WidgetTheme;
 import com.cleanroommc.modularui.utils.GlStateManager;
 import com.cleanroommc.modularui.utils.item.PlayerInvWrapper;
 import com.cleanroommc.modularui.utils.item.PlayerMainInvWrapper;
-import com.cleanroommc.modularui.value.StringValue;
 import com.cleanroommc.modularui.value.sync.ItemSlotSH;
 import com.cleanroommc.modularui.value.sync.PanelSyncManager;
 import com.cleanroommc.modularui.value.sync.SyncHandler;
-import com.cleanroommc.modularui.widget.ParentWidget;
 import com.cleanroommc.modularui.widget.Widget;
 import com.cleanroommc.modularui.widgets.ButtonWidget;
 import com.cleanroommc.modularui.widgets.layout.Column;
@@ -51,7 +43,6 @@ import com.cleanroommc.modularui.widgets.layout.Row;
 import com.cleanroommc.modularui.widgets.slot.ItemSlot;
 import com.cleanroommc.modularui.widgets.slot.ModularSlot;
 import com.cleanroommc.modularui.widgets.slot.SlotGroup;
-import com.cleanroommc.modularui.widgets.textfield.TextFieldWidget;
 
 import ruiseki.okbackpack.Reference;
 import ruiseki.okbackpack.api.IStorageContainer;
@@ -87,12 +78,7 @@ import ruiseki.okbackpack.client.gui.widget.TileWidget;
 import ruiseki.okbackpack.client.gui.widget.updateGroup.UpgradeSlotGroupWidget;
 import ruiseki.okbackpack.client.gui.widget.updateGroup.UpgradeSlotUpdateGroup;
 import ruiseki.okbackpack.client.gui.widget.upgrade.ExpandedTabWidget;
-import ruiseki.okbackpack.common.entity.properties.BackpackProperty;
 import ruiseki.okbackpack.common.helpers.BackpackInventoryHelpers;
-import ruiseki.okbackpack.common.helpers.BackpackJsonReader;
-import ruiseki.okbackpack.common.helpers.BackpackJsonWriter;
-import ruiseki.okbackpack.common.helpers.BackpackMaterial;
-import ruiseki.okbackpack.common.helpers.BackpackSettingsTemplate;
 import ruiseki.okbackpack.common.item.crafting.CraftingUpgradeWrapper;
 import ruiseki.okcore.helper.ItemStackHelpers;
 import ruiseki.okcore.helper.LangHelpers;
@@ -111,7 +97,6 @@ public class BackpackPanel extends ModularPanel implements IStoragePanel<Backpac
     public static final int ERROR_BORDER_COLOR = 0xFFB02E26;
     public static final int ERROR_TEXT_COLOR = 0xB02E26;
     public static final int ERROR_DISPLAY_TICKS = 60;
-    private static final File SETTINGS_TEMPLATE_DIR = new File("config/" + Reference.MOD_ID + "/dump");
 
     private static final List<CyclicVariantButtonWidget.Variant> SORT_TYPE_VARIANTS = Arrays.asList(
         new CyclicVariantButtonWidget.Variant(
@@ -152,26 +137,9 @@ public class BackpackPanel extends ModularPanel implements IStoragePanel<Backpac
     public boolean isSortingSettingTabOpened = false;
     public boolean isResetOpenedTabs = false;
 
-    private final List<String> availableSettingsFiles = new ArrayList<>();
-    private final StringValue settingsInputValue = new StringValue("");
-    private SettingsInputMode activeSettingsInput = SettingsInputMode.NONE;
-    private TextFieldWidget settingsInputField;
-    private ButtonWidget<?> saveButton;
-    private ButtonWidget<?> exportButton;
-    private ParentWidget<?> settingsButtonContainer;
-    private int selectedSettingsPresetIndex = 0;
-    private int selectedImportFileIndex = 0;
-    private int settingsTemplateRefreshTicks = 0;
-
     @Nullable
     public UpgradeSlotChangeResult activeError;
     public float activeErrorSetTick;
-
-    private enum SettingsInputMode {
-        NONE,
-        SAVE_PRESET,
-        EXPORT_FILE
-    }
 
     public BackpackPanel(EntityPlayer player, TileEntity tile, PanelSyncManager syncManager, UISettings settings,
         BackpackWrapper wrapper, int width, Integer backpackSlotIndex) {
@@ -263,11 +231,6 @@ public class BackpackPanel extends ModularPanel implements IStoragePanel<Backpac
     public void onInit() {
         super.onInit();
         updateListHeight();
-        if (settingsButtonContainer != null) {
-            getContext().getUISettings()
-                .getRecipeViewerSettings()
-                .addExclusionArea(settingsButtonContainer);
-        }
     }
 
     @Override
@@ -455,229 +418,6 @@ public class BackpackPanel extends ModularPanel implements IStoragePanel<Backpac
             .child(sleepButton);
     }
 
-    public void addSettingsTemplateButtons() {
-        refreshAvailableSettingsFiles();
-
-        settingsButtonContainer = new ParentWidget<>();
-        settingsButtonContainer.bottom(11)
-            .right(-21)
-            .size(18, 90)
-            .setEnabledIf(w -> settingPanel.isPanelOpen());
-
-        saveButton = new ButtonWidget() {
-
-            @Override
-            public boolean onMouseScroll(UpOrDown scrollDirection, int amount) {
-                cycleSettingsPreset(scrollDirection == UpOrDown.UP ? -1 : 1);
-                return true;
-            }
-        };
-        saveButton.bottom(72)
-            .left(0)
-            .size(18);
-        saveButton.overlay(OKBGuiTextures.SAVE_TEMPLATE_ICON);
-        saveButton.onMousePressed(mouseButton -> {
-            if (mouseButton == 0) {
-                if (activeSettingsInput == SettingsInputMode.SAVE_PRESET) {
-                    submitSettingsInput();
-                } else if (isSelectedPresetUnnamed()) {
-                    openSettingsInput(SettingsInputMode.SAVE_PRESET);
-                    alignInputFieldToButton(saveButton);
-                } else {
-                    saveCurrentSettingsPreset(getSelectedSettingsPresetEditableName());
-                }
-                return true;
-            }
-            return false;
-        });
-        saveButton.tooltipAutoUpdate(true)
-            .tooltipDynamic(tooltip -> {
-                String presetLabel;
-                if (selectedSettingsPresetIndex >= wrapper.getSettingsPresetCount()) {
-                    presetLabel = EnumChatFormatting.GREEN
-                        + LangHelpers.localize("gui.backpack.settings_preset_save_custom_name")
-                        + EnumChatFormatting.RESET;
-                } else {
-                    presetLabel = EnumChatFormatting.GREEN + getSelectedSettingsPresetLabel()
-                        + EnumChatFormatting.RESET;
-                }
-                tooltip.addLine(IKey.lang("gui.backpack.settings_preset_save", presetLabel))
-                    .addLine(
-                        IKey.lang("gui.backpack.settings_preset_save_controls")
-                            .style(IKey.GRAY, IKey.ITALIC))
-                    .pos(RichTooltip.Pos.NEXT_TO_MOUSE);
-            });
-
-        ButtonWidget<?> loadButton = new ButtonWidget() {
-
-            @Override
-            public boolean onMouseScroll(UpOrDown scrollDirection, int amount) {
-                cycleExistingPreset(scrollDirection == UpOrDown.UP ? -1 : 1);
-                return true;
-            }
-        };
-        loadButton.bottom(54)
-            .left(0)
-            .size(18);
-        loadButton.overlay(OKBGuiTextures.LOAD_TEMPLATE_ICON);
-        loadButton.onMousePressed(mouseButton -> {
-            if (mouseButton == 0) {
-                loadSelectedSettingsPreset();
-                return true;
-            }
-            return false;
-        });
-        loadButton.tooltipAutoUpdate(true)
-            .tooltipDynamic(tooltip -> {
-                String presetLabel = EnumChatFormatting.GREEN + getSelectedSettingsPresetLabel()
-                    + EnumChatFormatting.RESET;
-                tooltip.addLine(IKey.lang("gui.backpack.settings_preset_load", presetLabel))
-                    .addLine(
-                        IKey.lang("gui.backpack.settings_preset_load_controls")
-                            .style(IKey.GRAY, IKey.ITALIC))
-                    .pos(RichTooltip.Pos.NEXT_TO_MOUSE);
-            });
-
-        ButtonWidget<?> deleteButton = new ButtonWidget() {
-
-            @Override
-            public boolean onMouseScroll(UpOrDown scrollDirection, int amount) {
-                cycleExistingPreset(scrollDirection == UpOrDown.UP ? -1 : 1);
-                return true;
-            }
-        };
-        deleteButton.bottom(36)
-            .left(0)
-            .size(18);
-        deleteButton.overlay(OKBGuiTextures.DELETE_TEMPLATE_ICON);
-        deleteButton.onMousePressed(mouseButton -> {
-            if (mouseButton == 0) {
-                deleteSelectedSettingsPreset();
-                return true;
-            }
-            return false;
-        });
-        deleteButton.tooltipAutoUpdate(true)
-            .tooltipDynamic(tooltip -> {
-                if (selectedSettingsPresetIndex >= wrapper.getSettingsPresetCount()) {
-                    tooltip.addLine(IKey.lang("gui.backpack.settings_preset_delete_controls")
-                        .style(IKey.GRAY, IKey.ITALIC))
-                        .pos(RichTooltip.Pos.NEXT_TO_MOUSE);
-                    return;
-                }
-                String presetLabel = EnumChatFormatting.GREEN + getSelectedSettingsPresetLabel()
-                    + EnumChatFormatting.RESET;
-                tooltip.addLine(IKey.lang("gui.backpack.settings_preset_delete", presetLabel))
-                    .addLine(
-                        IKey.lang("gui.backpack.settings_preset_delete_controls")
-                            .style(IKey.GRAY, IKey.ITALIC))
-                    .pos(RichTooltip.Pos.NEXT_TO_MOUSE);
-            });
-
-        exportButton = new ButtonWidget<>().bottom(18)
-            .left(0)
-            .size(18)
-            .overlay(OKBGuiTextures.EXPORT_TEMPLATE_ICON)
-            .onMousePressed(mouseButton -> {
-                if (mouseButton == 0) {
-                    if (activeSettingsInput == SettingsInputMode.EXPORT_FILE) {
-                        submitSettingsInput();
-                    } else {
-                        exportCurrentSettingsToFile(getSuggestedExportName());
-                    }
-                    return true;
-                }
-                return false;
-            })
-            .tooltipAutoUpdate(true)
-            .tooltipDynamic(tooltip -> {
-                String exportName = getSuggestedExportName();
-                String displayName;
-                if (exportName.isEmpty()) {
-                    displayName = EnumChatFormatting.GREEN
-                        + LangHelpers.localize("gui.backpack.settings_preset_export_enter_name")
-                        + EnumChatFormatting.RESET;
-                } else {
-                    displayName = EnumChatFormatting.GREEN + exportName + EnumChatFormatting.RESET;
-                }
-                tooltip.addLine(IKey.lang("gui.backpack.settings_preset_export", displayName))
-                    .addLine(
-                        IKey.lang("gui.backpack.settings_preset_export_info")
-                            .style(IKey.GRAY, IKey.ITALIC))
-                    .pos(RichTooltip.Pos.NEXT_TO_MOUSE);
-            });
-
-        ButtonWidget<?> importButton = new ButtonWidget() {
-
-            @Override
-            public boolean onMouseScroll(UpOrDown scrollDirection, int amount) {
-                cycleImportFile(scrollDirection == UpOrDown.UP ? -1 : 1);
-                return true;
-            }
-        };
-        importButton.bottom(0)
-            .left(0)
-            .size(18);
-        importButton.overlay(OKBGuiTextures.IMPORT_TEMPLATE_ICON);
-        importButton.onMousePressed(mouseButton -> {
-            if (mouseButton == 0) {
-                importSelectedSettingsFile();
-                return true;
-            }
-            return false;
-        });
-        importButton.tooltipAutoUpdate(true)
-            .tooltipDynamic(tooltip -> {
-                if (availableSettingsFiles.isEmpty()) {
-                    tooltip.addLine(IKey.lang("gui.backpack.settings_preset_no_files"));
-                } else {
-                    String importFile = EnumChatFormatting.GREEN + getSelectedImportFileName()
-                        + EnumChatFormatting.RESET;
-                    tooltip.addLine(IKey.lang("gui.backpack.settings_preset_import", importFile))
-                        .addLine(
-                            IKey.lang("gui.backpack.settings_preset_import_controls")
-                                .style(IKey.GRAY, IKey.ITALIC));
-                }
-                tooltip.pos(RichTooltip.Pos.NEXT_TO_MOUSE);
-            });
-
-        settingsInputField = new TextFieldWidget() {
-
-            @Override
-            public void setText(@NotNull String text) {
-                super.setText(text);
-                handler.setCursor(0, text.length(), false);
-            }
-
-            @Override
-            public @NotNull Interactable.Result onKeyPressed(char character, int keyCode) {
-                if (keyCode == Keyboard.KEY_RETURN || keyCode == Keyboard.KEY_NUMPADENTER) {
-                    submitSettingsInput();
-                    return Interactable.Result.SUCCESS;
-                }
-                if (keyCode == Keyboard.KEY_ESCAPE) {
-                    closeSettingsInput();
-                    return Interactable.Result.SUCCESS;
-                }
-                return super.onKeyPressed(character, keyCode);
-            }
-        }.value(settingsInputValue)
-            .setMaxLength(128)
-            .background(OKBGuiTextures.ANVIL_TEXT_FIELD_ENABLED)
-            .bottom(101)
-            .left(0)
-            .size(72, 16)
-            .setEnabledIf(widget -> settingPanel.isPanelOpen() && activeSettingsInput != SettingsInputMode.NONE);
-
-        settingsButtonContainer.child(saveButton)
-            .child(loadButton)
-            .child(deleteButton)
-            .child(exportButton)
-            .child(importButton);
-
-        child(settingsButtonContainer).child(settingsInputField);
-    }
-
     public void addMainWidget() {
         slotRow = (Row) new Row().coverChildren()
             .top(18)
@@ -824,27 +564,6 @@ public class BackpackPanel extends ModularPanel implements IStoragePanel<Backpac
     @Override
     public void onUpdate() {
         super.onUpdate();
-
-        if (!settingPanel.isPanelOpen() && activeSettingsInput != SettingsInputMode.NONE) {
-            closeSettingsInput();
-        }
-
-        if (settingPanel.isPanelOpen() && saveButton != null && exportButton != null) {
-            if (exportButton.isHovering()) {
-                if (activeSettingsInput != SettingsInputMode.EXPORT_FILE) {
-                    openSettingsInput(SettingsInputMode.EXPORT_FILE);
-                    alignInputFieldToButton(exportButton);
-                }
-            } else if (activeSettingsInput != SettingsInputMode.NONE
-                && (settingsInputField == null || !settingsInputField.isHovering())) {
-                    closeSettingsInput();
-                }
-        }
-
-        if (++settingsTemplateRefreshTicks >= 20) {
-            settingsTemplateRefreshTicks = 0;
-            refreshAvailableSettingsFiles();
-        }
     }
 
     @SuppressWarnings({ "unchecked", "rawtypes" })
@@ -1075,268 +794,6 @@ public class BackpackPanel extends ModularPanel implements IStoragePanel<Backpac
         backpackSyncHandler.syncToServer(
             BackpackSH.getId(BackpackSHRegisters.UPDATE_SEARCH_PHRASE),
             buffer -> buffer.writeStringToBuffer(""));
-    }
-
-    private void refreshAvailableSettingsFiles() {
-        if (!SETTINGS_TEMPLATE_DIR.exists()) {
-            SETTINGS_TEMPLATE_DIR.mkdirs();
-        }
-
-        availableSettingsFiles.clear();
-        File[] files = SETTINGS_TEMPLATE_DIR.listFiles((dir, name) -> name.endsWith(".json"));
-        if (files != null) {
-            for (File file : files) {
-                String name = file.getName();
-                availableSettingsFiles.add(name.substring(0, name.length() - 5));
-            }
-        }
-        Collections.sort(availableSettingsFiles);
-        normalizeSelectedIndexes();
-    }
-
-    private void normalizeSelectedIndexes() {
-        int presetCount = Math.max(1, wrapper.getSettingsPresetCount() + 1);
-        selectedSettingsPresetIndex = Math.max(0, Math.min(selectedSettingsPresetIndex, presetCount - 1));
-
-        if (availableSettingsFiles.isEmpty()) {
-            selectedImportFileIndex = 0;
-        } else {
-            selectedImportFileIndex = Math.floorMod(selectedImportFileIndex, availableSettingsFiles.size());
-        }
-    }
-
-    private void cycleSettingsPreset(int delta) {
-        int presetCount = Math.max(1, wrapper.getSettingsPresetCount() + 1);
-        selectedSettingsPresetIndex = Math.floorMod(selectedSettingsPresetIndex + delta, presetCount);
-    }
-
-    private void cycleExistingPreset(int delta) {
-        int presetCount = wrapper.getSettingsPresetCount();
-        if (presetCount <= 0) return;
-        selectedSettingsPresetIndex = Math.floorMod(selectedSettingsPresetIndex + delta, presetCount);
-    }
-
-    private void cycleImportFile(int delta) {
-        refreshAvailableSettingsFiles();
-        if (!availableSettingsFiles.isEmpty()) {
-            selectedImportFileIndex = Math.floorMod(selectedImportFileIndex + delta, availableSettingsFiles.size());
-        }
-    }
-
-    private void openSettingsInput(SettingsInputMode mode) {
-        if (activeSettingsInput == mode) {
-            closeSettingsInput();
-            return;
-        }
-
-        activeSettingsInput = mode;
-        String text = mode == SettingsInputMode.SAVE_PRESET ? getSelectedSettingsPresetEditableName()
-            : getSuggestedExportName();
-        settingsInputValue.setStringValue(text);
-        settingsInputField.setText(text);
-    }
-
-    private void closeSettingsInput() {
-        activeSettingsInput = SettingsInputMode.NONE;
-        settingsInputValue.setStringValue("");
-        if (settingsInputField != null) {
-            settingsInputField.setText("");
-        }
-    }
-
-    private boolean isSelectedPresetUnnamed() {
-        if (selectedSettingsPresetIndex >= wrapper.getSettingsPresetCount()) {
-            return true;
-        }
-        String name = wrapper.getSettingsPresetName(selectedSettingsPresetIndex);
-        return name == null || name.isEmpty();
-    }
-
-    private void alignInputFieldToButton(ButtonWidget<?> button) {
-        if (settingsInputField == null || button == null) return;
-        int inputWidth = settingsInputField.getArea().width;
-        int buttonAreaX = button.getArea().x;
-        int leftX = buttonAreaX - 1 - inputWidth;
-        settingsInputField.pos(leftX - getArea().x, button.getArea().y - getArea().y);
-    }
-
-    private void submitSettingsInput() {
-        String input = settingsInputField == null ? settingsInputValue.getStringValue() : settingsInputField.getText();
-        if (input == null) {
-            input = "";
-        }
-
-        if (activeSettingsInput == SettingsInputMode.SAVE_PRESET) {
-            saveCurrentSettingsPreset(input.trim());
-        } else if (activeSettingsInput == SettingsInputMode.EXPORT_FILE) {
-            exportCurrentSettingsToFile(input.trim());
-        }
-
-        closeSettingsInput();
-    }
-
-    private void saveCurrentSettingsPreset(String name) {
-        String presetName = name.isEmpty() ? getSelectedSettingsPresetEditableName() : name;
-        wrapper.saveSettingsPreset(selectedSettingsPresetIndex, presetName);
-        backpackSyncHandler.syncToServer(BackpackSH.getId(BackpackSHRegisters.UPDATE_SAVE_SETTINGS_PRESET), buffer -> {
-            buffer.writeInt(selectedSettingsPresetIndex);
-            buffer.writeStringToBuffer(presetName == null ? "" : presetName);
-        });
-    }
-
-    private void loadSelectedSettingsPreset() {
-        if (selectedSettingsPresetIndex >= wrapper.getSettingsPresetCount()) {
-            return;
-        }
-
-        if (wrapper.loadSettingsPreset(selectedSettingsPresetIndex)) {
-            syncLocalPlayerSettingsFromWrapper();
-            updateUpgradeWidgets();
-            backpackSyncHandler.syncToServer(
-                BackpackSH.getId(BackpackSHRegisters.UPDATE_LOAD_SETTINGS_PRESET),
-                buffer -> buffer.writeInt(selectedSettingsPresetIndex));
-        }
-    }
-
-    private void deleteSelectedSettingsPreset() {
-        if (selectedSettingsPresetIndex >= wrapper.getSettingsPresetCount()) {
-            return;
-        }
-
-        selectedSettingsPresetIndex = wrapper.deleteSettingsPreset(selectedSettingsPresetIndex);
-        backpackSyncHandler.syncToServer(
-            BackpackSH.getId(BackpackSHRegisters.UPDATE_DELETE_SETTINGS_PRESET),
-            buffer -> buffer.writeInt(selectedSettingsPresetIndex));
-
-        if (wrapper.getSettingsPresetCount() == 0) {
-            saveCurrentSettingsPreset(LangHelpers.localize("gui.backpack.settings_preset_new_slot"));
-        }
-
-        normalizeSelectedIndexes();
-    }
-
-    private void exportCurrentSettingsToFile(String input) {
-        refreshAvailableSettingsFiles();
-
-        String fileName = input.isEmpty() ? getSuggestedExportName() : input;
-        if (fileName.isEmpty()) {
-            fileName = "settings";
-        }
-
-        BackpackMaterial material = new BackpackMaterial();
-        material.setBackpackTier("Base");
-        material.setSettingsFromTemplate(BackpackSettingsTemplate.fromWrapper(wrapper));
-
-        File file = new File(SETTINGS_TEMPLATE_DIR, fileName + ".json");
-        try {
-            new BackpackJsonWriter(file).write(material);
-        } catch (IOException e) {
-            return;
-        }
-
-        refreshAvailableSettingsFiles();
-        selectedImportFileIndex = availableSettingsFiles.indexOf(fileName);
-        if (selectedImportFileIndex < 0) {
-            selectedImportFileIndex = 0;
-        }
-    }
-
-    private void importSelectedSettingsFile() {
-        refreshAvailableSettingsFiles();
-        if (availableSettingsFiles.isEmpty()) {
-            return;
-        }
-
-        String fileName = getSelectedImportFileName();
-        if (fileName.isEmpty()) {
-            return;
-        }
-
-        try {
-            BackpackMaterial material = new BackpackJsonReader(new File(SETTINGS_TEMPLATE_DIR, fileName + ".json"))
-                .read();
-            if (material == null || !material.hasSettings()) {
-                return;
-            }
-
-            String presetName = getUniqueImportedPresetName(fileName);
-            BackpackSettingsTemplate template = material.toSettingsTemplate(wrapper.getSlots());
-            selectedSettingsPresetIndex = wrapper.addSettingsPreset(presetName, template);
-            backpackSyncHandler
-                .syncToServer(BackpackSH.getId(BackpackSHRegisters.UPDATE_IMPORT_SETTINGS_PRESET), buffer -> {
-                    buffer.writeStringToBuffer(presetName);
-                    buffer.writeNBTTagCompoundToBuffer(template.serializeNBT());
-                });
-        } catch (IOException e) {
-            // ignore invalid files and keep current state unchanged
-        }
-    }
-
-    private void syncLocalPlayerSettingsFromWrapper() {
-        if (!wrapper.isUsePlayerSettings()) {
-            return;
-        }
-
-        BackpackProperty property = BackpackProperty.get(player);
-        if (property == null) {
-            return;
-        }
-
-        property.setLockBackpack(wrapper.isLockStorage());
-        property.setKeepTab(wrapper.isKeepTab());
-        property.setShiftClickIntoOpenTab(wrapper.isShiftClickIntoOpenTab());
-        property.setKeepSearchPhrase(wrapper.isKeepSearchPhrase());
-    }
-
-    private String getSelectedSettingsPresetEditableName() {
-        return selectedSettingsPresetIndex < wrapper.getSettingsPresetCount()
-            ? wrapper.getSettingsPresetName(selectedSettingsPresetIndex)
-            : "";
-    }
-
-    private String getSelectedSettingsPresetLabel() {
-        String name = getSelectedSettingsPresetEditableName();
-        if (selectedSettingsPresetIndex >= wrapper.getSettingsPresetCount()) {
-            return LangHelpers.localize("gui.backpack.settings_preset_new_slot");
-        }
-        return name == null || name.isEmpty() ? LangHelpers.localize("gui.backpack.settings_preset_unnamed") : name;
-    }
-
-    private String getSuggestedExportName() {
-        String name = getSelectedSettingsPresetEditableName();
-        return name == null || name.trim()
-            .isEmpty() ? "" : name.trim();
-    }
-
-    private String getSuggestedExportDisplayName() {
-        String name = getSuggestedExportName();
-        return name.isEmpty() ? LangHelpers.localize("gui.backpack.settings_preset_export_enter_name") : name;
-    }
-
-    private String getSelectedImportFileName() {
-        if (availableSettingsFiles.isEmpty()) {
-            return "";
-        }
-        return availableSettingsFiles.get(selectedImportFileIndex);
-    }
-
-    private String getUniqueImportedPresetName(String baseName) {
-        String candidate = (baseName == null || baseName.trim()
-            .isEmpty()) ? "imported" : baseName.trim();
-        List<String> existingNames = new ArrayList<>();
-        for (int i = 0; i < wrapper.getSettingsPresetCount(); i++) {
-            existingNames.add(wrapper.getSettingsPresetName(i));
-        }
-
-        if (!existingNames.contains(candidate)) {
-            return candidate;
-        }
-
-        int suffix = 2;
-        while (existingNames.contains(candidate + " (" + suffix + ")")) {
-            suffix++;
-        }
-        return candidate + " (" + suffix + ")";
     }
 
     public int getOpenCraftingUpgradeSlot() {
